Some more restructuring for testability and sensibility.
Lost Jane Austen novel: Tests & Testability, about a trio of sisters
whose father has recently died. The oldest one is a techwriter. The
middle one with the tempestuous personality is in QA. The youngest
spends her days hiding under a table in the library hacking on her IRC
bot project.